Commit - ajustes finais - buddy bridge - módulo de animais

+++ buddybridge-frontend/src/app/restrict/base/adoption-profile/component/adoption/adoption-profile.component.ts
@@ -5,22 +5,26 @@ import {ActivatedRoute, Router} from "@angular/router";
 import {MessageService} from "primeng/api";
 import {AdoptionService} from "../../shared/adoption.service";
 import {AdoptionProfileModel} from "../../model/AdoptionProfileModel";
+import { GrupoacessoserviceService } from '../../../grupo_acesso/service/grupoacessoservice.service';
+import { AcessoDTO } from '../../../grupo_acesso/model/acessoDTO';
 
 @Component({
   selector: 'app-adoption-profile',
-  standalone: true,
-  imports: [
-    AdoptionProfileListComponent
-  ],
   templateUrl: './adoption-profile.component.html',
   styleUrl: './adoption-profile.component.scss'
 })
 export class AdoptionProfileComponent {
   profiles!: AdoptionProfileModel[]
 
+  //Fazendo a solicitação de acessos - inicio
+  acessos: AcessoDTO[] = [];
+  displayModal: boolean = false;
+  //Fazendo a solicitação de acessos - fim
+
   constructor(
     private adoptionService: AdoptionService,
     private accountService: AccountService,
+    private grupoacessoserviceService: GrupoacessoserviceService,
     private router: Router,
     private route: ActivatedRoute,
     private messageService: MessageService
@@ -28,7 +32,26 @@ export class AdoptionProfileComponent {
     this.refresh();
   }
 
-  ngOnInit(): void { this.accountService.validarSessao(); }
+  ngOnInit(): void {
+    this.accountService.validarSessao();
+    this.grupoacessoserviceService.getAcessosParaTela('Cadastro de perfil de adoção').subscribe((acessos: AcessoDTO[]) => {
+      this.acessos = acessos;
+     });
+  }
+
+  openModal() {
+    this.displayModal = true;
+  }
+
+  solicitarAcesso(acesso: AcessoDTO) {
+    this.grupoacessoserviceService.solicitarAcesso(acesso.idAcesso).subscribe(() => {
+        this.messageService.add({ severity: 'success', summary: 'Solicitação feita com sucesso', detail: 'Aguarde o administrador' });
+      },
+      error => {
+        this.messageService.add({ severity: 'warning', summary: 'Atenção!', detail: 'Acesso já solicitado!' });
+      }
+    );
+  }
 
   refresh() {
     this.adoptionService.getAnimalsByProfileStatus().subscribe((data: AdoptionProfileModel[]) => {
@@ -37,18 +60,41 @@ export class AdoptionProfileComponent {
   }
 
   onAdd() {
-    this.router.navigate(['addperfil'], { relativeTo: this.route });
+    if (this.hasAccess('Cadastrar perfil de adoção')) {
+      this.router.navigate(['addperfil'], { relativeTo: this.route });
+    } else {
+      this.showAccessDeniedModal();
+    }
+
   }
 
   onEdit(idUser: any) {
-    this.router.navigate(['editperfil', idUser], { relativeTo: this.route });
+    if (this.hasAccess('Alterar perfil de adoção')) {
+      this.router.navigate(['editperfil', idUser], { relativeTo: this.route });
+    } else {
+      this.showAccessDeniedModal();
+    }
+
   }
 
   onRemove(idUser: any) {
-    this.adoptionService.deleteAdoptionProfile(idUser).subscribe(() =>{
-      this.messageService.add({ severity: 'success', summary: 'Sucesso', detail: 'Registro excluido com sucesso' });
-      window.location.reload();
-      this.router.navigateByUrl('/perfil-adocao')
-    })
+    if (this.hasAccess('Excluir perfil de adoção')) {
+      this.adoptionService.deleteAdoptionProfile(idUser).subscribe(() =>{
+        this.messageService.add({ severity: 'success', summary: 'Sucesso', detail: 'Registro excluido com sucesso' });
+        window.location.reload();
+        this.router.navigateByUrl('/perfil-adocao')
+      })
+    } else {
+      this.showAccessDeniedModal();
+    }
+  }
+
+  private hasAccess(descricaoAcesso: string): boolean {
+    return !this.acessos.some(acesso => acesso.descricaoAcesso === descricaoAcesso);
+  }
+
+  private showAccessDeniedModal() {
+    this.messageService.add({ severity: 'warning', summary: 'Atenção!', detail: 'Você não possui acesso a esta funcionalidade.' });
+    this.openModal();
   }
 }
